requirements.txt
python-magic
requests
docopt